requirements.txt
streamlit
pandas
numpy
requests
python-dotenv
plotly